[FIX] Refactor code, fix bug with color selection, add button styles
