+++ src/styles/style.css
@@ -56,12 +56,20 @@ form#link-form {
       width: 35%;
       background: white;
       color: rgba(51, 51, 51, 1);
+
+      &:hover {
+        background-color: rgb(223, 238, 255);
+      }
     }
 
     button.dload-btn {
       width: 65%;
       background: var(--primary-color);
       color: white;
+
+      &:hover {
+        background-color: rgb(54, 144, 218);
+      }
     }
   }
 
@@ -69,6 +77,10 @@ form#link-form {
     width: 100%;
     background: var(--secondary-color);
     color: white;
+
+    &:hover {
+      background-color: rgb(150, 100, 219);
+    }
   }
 
   textarea#link-input {
@@ -105,6 +117,10 @@ footer {
       cursor: pointer;
       background-color: white;
 
+      &:hover {
+        background-color: rgb(216, 231, 255);
+      }
+
       p {
         margin: 10px 0;
         position: relative;
@@ -127,21 +143,25 @@ footer {
     p.year {
       position: absolute;
       bottom: -10px;
-      right: 28px;
+      right: 25px;
       z-index: 2;
     }
 
-    p.settings {
+    p.website {
       position: absolute;
       bottom: -10px;
-      left: 28px;
+      left: 25px;
       z-index: 2;
-      cursor: pointer;
 
-      .settings-icon {
+      a {
+        text-decoration: none;
+        color: inherit;
+      }
+
+      .website-icon {
         font-size: 16px;
         position: relative;
-        bottom: -2px;
+        bottom: -3px;
       }
     }
   }
@@ -214,5 +234,9 @@ section#color-page {
     border-radius: 10px;
     cursor: pointer;
     font-size: 16px;
+
+    &:hover {
+      background-color: rgb(54, 144, 218);
+    }
   }
 }
